requirements.txt
PyPDF2>=3.0.0
pyyaml>=6.0
tqdm>=4.65.0
pycryptodome
